+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 V2.0.5
 - Pass PIP_INDEX_URL and PIP_PRE to the docker containers that run the tests.
+- Fixed dict value null conversion to None in Jinja template (#97)
 
 V2.0.4
 - Add support to run the tests on centos8